UPSTREAM: soc/intel/tigerlake: Provide SoundWire controller properties
The Intel Tigerlake SoundWire controller has 4 master links which
are configured differently depending on the external crystal oscillator
which is connected to the PCH.

This function will read the PCH PMC EPOC register to determine the
frequency and then fill out the master link entries with the correct
table values.

The frequency is also provided directly in a custom "ip-clock" property
which will be added to the link descriptor and passed to the OS driver
so it can know the clock rate of the master.

BUG=b:146482091

Change-Id: Icb0a99f829ebfe74d53fb5a803f25ae46d9ae3d3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2d0655008fa85d19c0c9e95dec9b26522fe2951f
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Change-Id: I98b7df21210c29cd8defeff648f2c2207d629295
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40889
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2216002
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>