UPSTREAM: libpayload: xhci: Do not memcpy registers
memcpy() is meant to be used on normal memory and often implemented with
architecture-specific optimizations to make that as performant as
possible. MMIO registers often have special access restrictions that may
be incompatible with whatever memcpy() does. For example, on arm64 it
uses the LDP (load pair) to load 16 bytes at a time, which makes 4-byte
MMIO registers unhappy.

This patch removes the caching of the XHCI capreg registers and changes
it back to a pointer. The CAP_GET() macro is still accessing a full
(non-bitfield) uint32_t at the end so this should still generate a
4-byte access (which was the goal of the original change in CB:39838).

BUG=none
BRANCH=none
TEST=none

Change-Id: I0110f5cf8cc13ac98dee3eaef6e6facd9822a2a8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6c1a669b444fc7e6d7542ff910deca1a606d4c29
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Id058c8813087a8e8cb85f570399e07fb8a597108
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/40895
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2175818
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>