UPSTREAM: superio/winbond/w83977tf: Only list IR logical device if needed
Really old versions of W83977TF Super I/O had an IR logical device, but
is no longer the case. It does not exist in the newer W83977EF version,
installed in some Asus P2B family boards, and served by this same code.

Add a config option on the off chance we may see board with it (as if
we would) and don't include this device unless it is set. Saves us from
the need to declare a not-present device off and/or extraneous PNP
device errors about a not-present device.

BUG=none
BRANCH=none
TEST=none

Change-Id: I00507d043f72a934c30806e0ee2bf87674a7c595
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f5472a10c6fc3f0c452c514c7a23b7046e42ac55
Original-Change-Id: I761ebc41f1735a03e768339a68ca139167edc095
Original-Signed-off-by: Keith Hui <buurin@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41004
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2193894
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>