UPSTREAM: security/vboot: Add option to run verstage before bootblock
For AMD's family 17h, verstage can run as a userspace app in the PSP
before the X86 is released. The flags for this have been made generic
to support any other future systems that might run verstage before
the main processor starts.

Although an attempt has been made to make things somewhat generic,
since this is the first and currently only chip to support verstage
before bootblock, there are a number of options which might ultimately
be needed which have currently been left out for simplicity.  Examples
of this are:
- PCI is not currently supported - this is currently just a given
instead of making a separate Kconfig option for it.
- The PSP uses an ARM v7 processor, so that's the only processor that
is getting updated for the verstage-before-bootblock option.

BUG=b:158124527
TEST=Build with following patches

Change-Id: I9a1695a9e8c6595902e179c35a2fc472a933b97c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8a3a3c820b8f8a8d357bf54fb7532ad1ae1ba270
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Change-Id: I4849777cb7ba9f90fe8428b82c21884d1e662b96
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41814
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2246030
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>