UPSTREAM: memrange: Break early from memranges_find_entry if limit is crossed
This change updates memranges_find_entry() to break and return early
if the end address of the hole within the current range entry crosses
the requested limit. This is because all range entries and maintained
in increasing order and so none of the following range entries can
satisfy the given request.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2649717ae0e4aa36bab5c5175beddea3632c6985
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f8f5650873451f64eab84d4b20248419bc19a0f5
Original-Change-Id: I14e03946ddbbb5d254b23e9a9917da42960313a6
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41104
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2190947
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>