UPSTREAM: cbfstool: Support CONFIG_CBFS_VERIFICATION and metadata hash anchor
This patch adds support for the new CONFIG_CBFS_VERIFICATION feature to
cbfstool. When CBFS verification is enabled, cbfstool must automatically
add a hash attribute to every CBFS file it adds (with a handful of
exceptions like bootblock and "header" pseudofiles that are never read
by coreboot code itself). It must also automatically update the metadata
hash that is embedded in the bootblock code. It will automatically find
the metadata hash by scanning the bootblock for its magic number and use
its presence to auto-detect whether CBFS verification is enabled for an
image (and which hash algorithm to use).

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 4bfbabdb54ed6a56bdfa9e703b49f4ed7d9a6acc
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I61a84add8654f60c683ef213b844a11b145a5cb7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41121
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: If855268899ee813ebfa9120eb31d82e2cc842775
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2759207
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>