UPSTREAM: mb/asus/p8z77-m_pro: Remove PS/2 keyboard & mouse duplicate
PS/2 keyboard and mouse devices are declared twice in the DSDT, once
in mainboard and once in southbridge. It would appear in Windows
Device Manager as two PS/2 keyboards and two PS/2 mouses, all with
resource conflicts. This change drops the declaration from mainboard.

The issue was discovered when this setup was copied for p8z77-m and
being boot tested.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia7ca2b6e61be3e94c5b31d6d89cc284b2635dff4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 698647a8c0195e7920b6bf6614c81a3085af7a38
Original-Change-Id: I746a960aaf3992acbcb6a7364641fc4fd12002d2
Original-Signed-off-by: Keith Hui <buurin@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/41225
Original-Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2283116
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>